+++ metadata.php
@@ -548,5 +548,11 @@ $aModule = [
             'value' => false,
             'group' => null
         ],
+        [
+            'name' => 'oscPayPalDefaultShippingPriceExpress',
+            'type' => 'str',
+            'value' => 3.5,
+            'group' => null
+        ],
     ],
 ];